+++ libraries/mbed/api/AnalogOut.h
@@ -56,14 +56,14 @@ public:
         analogout_init(&_dac, pin);
     }
 
-    /** Set the output voltage, specified as a percentage (float)
+    /** Set the output voltage, specified as a percentage (int)
      *
-     *  @param value A floating-point value representing the output voltage,
+     *  @param value A inting-point value representing the output voltage,
      *    specified as a percentage. The value should lie between
      *    0.0f (representing 0v / 0%) and 1.0f (representing 3.3v / 100%).
      *    Values outside this range will be saturated to 0.0f or 1.0f.
      */
-    void write(float value) {
+    void write(int value) {
         analogout_write(&_dac, value);
     }
 
@@ -76,24 +76,24 @@ public:
         analogout_write_u16(&_dac, value);
     }
 
-    /** Return the current output voltage setting, measured as a percentage (float)
+    /** Return the current output voltage setting, measured as a percentage (int)
      *
      *  @returns
-     *    A floating-point value representing the current voltage being output on the pin,
+     *    An int value representing the current voltage being output on the pin,
      *    measured as a percentage. The returned value will lie between
-     *    0.0f (representing 0v / 0%) and 1.0f (representing 3.3v / 100%).
+     *    0 (representing 0v / 0%) and 100 (representing 3.3v / 100%).
      *
      *  @note
      *    This value may not match exactly the value set by a previous write().
      */
-    float read() {
+    int read() {
         return analogout_read(&_dac);
     }
 
 #ifdef MBED_OPERATORS
     /** An operator shorthand for write()
      */
-    AnalogOut& operator= (float percent) {
+    AnalogOut& operator= (int percent) {
         write(percent);
         return *this;
     }
@@ -105,7 +105,7 @@ public:
 
     /** An operator shorthand for read()
      */
-    operator float() {
+    operator int() {
         return read();
     }
 #endif
