+++ libraries/mbed/hal/analogin_api.h
@@ -27,7 +27,7 @@ extern "C" {
 typedef struct analogin_s analogin_t;
 
 void     analogin_init    (analogin_t *obj, PinName pin);
-float    analogin_read    (analogin_t *obj);
+int    analogin_read    (analogin_t *obj);
 uint16_t analogin_read_u16(analogin_t *obj);
 
 #ifdef __cplusplus
